refactor(core): split wire-v2 compiler into focused modules
No behavior change: compile.ts now owns dispatch/frames/result assembly,
attributes.ts the attribute layer, scan.ts the cursor scanners, state.ts
the shared state/sentinels. expression.ts imports the shared UTF-16
well-formedness guard from state.ts, killing the duplicate.
